+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 ENV["RAILS_ENV"] = "test"
 
-require "test/rails_app/config/environment"
+require "rails_app/config/environment"
 require "rails/test_help"
 
 require 'ajax_pagination'